requirements.txt
livekit-agents[google]>=1.2.0
livekit-plugins-google>=1.2.18
google-genai>=1.47.0
python-dotenv>=1.0.1
feedparser>=6.0.10
playwright>=1.48.0
